generando las entradas junto con la realizacion de espectáculos

+++ src/main/java/com/grupo6/persistence/repository/EspectaculoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.grupo6.persistence.model.Espectaculo;
 
 @Repository
-public interface EspectaculoRepository extends BaseRepository <Espectaculo, Integer>{
+public interface EspectaculoRepository extends BaseRepository <Espectaculo, Long>{
 
 	@EntityGraph("Espectaculo.Full")
 	@Override
